move generic form into its own component
